Move events handling to offline plugin
