+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/plugin/state/StateRegistry.kt
@@ -60,6 +60,8 @@ public class StateRegistry internal constructor(
         }
     }
 
+    internal fun getActiveChannelStates(): List<ChannelState> = channels.values.toList()
+
     /** Clear state of all state objects. */
     public fun clear() {
         queryChannels.clear()
